Choose movie colors (#1)
Refactor BWColor to ReversibleColor.

Create a dual color picker based on the flet-contrib widget ; place it 
below script area (outside of to-be menu cards, for direct access ; this 
will make sense when saving settings, where the only thing you want to 
change are probably script and colors).

To fill the void, add stand-in for future movie preview, to the right.

Place generate area way at the bottom, where it should be.