First verson of video preview (#10)
Never get too pessimistic around midnight :)

This works, but it's *ungodly* slow. Every update to the preview takes
around ~3 seconds, which of course, only gets worse for the actual movie
rendering (~1m40s for a 17s clip ?!).

On short renders (previews), the bottleneck is writing the frames to
.tmp. Piping them would be faster, but does this package of ffmpeg
support that, or am I going to have to switch dependencies..?