Introduce flash_duration parameter
Defaults to 2 (one frame in main color, one in inverse color). 

Also change default phrase_duration to 15 flashes instead of 3 (meaning 
each phrase is on screen for a duration of 30 frames or one second by 
default).

Misc : add calculate_movie_length() which, for now, has no use (it would 
have been used in video_preview.seek(), but is not)