Add dialog-contained larger script window
This allows the user to edit the script in a more comfortable space than 
the restrained box of the main window.

The script window is called forth by an expand icon placed as a prefix 
to the main window script field. Icon alignment in text fields are a 
pain, though, because this framework is bad.

Otherwise, the only big change is that script_field now has an on_change 
aside from GenerateArea's operations.