First version of font selection
Using a flet SearchBar, looking through a list compiled by the 
Matplotlib font manager. 

There are no major bugs, but a number of small inconveniences : 

- Matplotlib is a very heavy dependency just for this functionality, but 
extracting the font manager appears non-trivial, as it contains imports 
to other Matplotlib content.
- The search bar expands above its initial location when opened, for no 
apparent reason. 
- Once an option is chosen, there is no way to simply remove focus from 
the search bar, and so we instead focus the font size input. 
- This also means that, when closing the video options pane, the font 
size retains focus, because there is no apparent way to blur a control.