Add _tmpdir_scope() function and use it in place of TEMP_OUTPUT_PATH
On top of readability, this allows for the tmpdir to be located in RAM 
filesystems on most platforms. However, in the case of an SSD main disk, 
there is no performance gain at all. Still, it is preferable in the 
general case.

Misc : remove .tmp/ from gitignore now, since it will no longer be 
created :)