Detailed progress info during render (#27)
The progress ring has been replaced by a full fledged progress bar. 
Surrounding it are completion %, time remaining, frame number and render 
FPS. Additionally, the progress percentage is also shown in the app 
title.

To transmit this information, a generic Unix stream socket (AF_UNIX type 
stream, also available in Windows since 2017, so cross-platform) is used 
as the output for the ffmpeg -progress logs. The socket is read by small 
buffers in a separate thread, and the info is sent on the page pubsub.

To use the pubsub in this way, the update_preview message had to be 
moved to a specific topic. The progress message is on its own topic as 
well.