Long press on NumberInput buttons (#30)
This is accomplished through GestureDetectors, which launch a thread 
running the increment function on loop until the button is released.

on_change() also works differently now, and only occurs after a long 
time ; this should mostly avoid the value changing from under the user's 
feet as they edit the field manually. 

In addition to this, the value is immediately updated when the field is 
blurred. 

The clamping and int conversion operation have been moved into a 
setter-like helper function, which is the only option available since 
setters for fields not decorated with @property cannot be implemented 
without infinite recursion.

There is one major inconvenience with this implementation : if the 
button is pressed, but the change doesn't immediately happen due to 
timing circumstances, then the user lets go and tries pressing again, 
the IO lock is taken, and so the increment waits until the preview from 
the first tap is rendered, after which the field brutally increments far 
beyond the desired value. Look into fixing this.

Misc : the width of the control has been slightly increased, as the 
GestureDetectors seem to take up virtual width and would prevent the 
one-line display of three-digit numbers (???)