Auto font size (#5)
MovieOptionsWrapper now carries a bool for auto font size. It will, in
the future, be the designated spot for such options that don't pertain
to the parameters of the video itself, but to the current app settings.

A new pubsub channel has been added, complete with the creation of two
separate methods in the wrapper for preview and font size messages.

main.py has been reorganized into component initialization, pubsub
subscriptions and page building successively, with a handler added for
font size that calls the previously written Movie method.

A switch was then added to VideoSection to enable or disable auto font
size. The associated number input is disabled when the font size is set
to auto.

Font size is recalculated upon changes to the script, capitalization,
font, resolution, or border.