Allow using mouse wheel in NumberInput (#40)
The mouse wheel can now be used to increment or decrement the value of a 
NumberInput control.

Several solutions were considered for this feature.

The first idea was to include this GestureDetector inside the 
NumberInput component. This, however, would have required making said 
component inherit from GestureDetector or Container, which would have 
severely complicated its internal logic.

The next idea was to create a NumberInputWrapper enclosing the control 
in a GestureDetector. However, again, the transmission of events and 
values across that barrier would have been cumbersome.

In line with the support of keyboard events, the next idea was to create 
a GestureDetector at the root Page level. Unfortunately, the 
GestureDetector ended up overshadowed by the VideoSection's edit window, 
and thus detected nothing.

The compromise that was finally reached was to define a handler similar 
to that of keyboard events at the page level, but to localize the 
GestureDetector in the VideoSection, wrapping the contents of its edit 
window inside it. The handler takes a "page" argument to bridge the two.